users mangement

+++ MyCMS/Admin/Role.aspx.cs
@@ -26,41 +26,51 @@ namespace MyCMS.Admin
 
         protected void gvRoles_RowCommand(object sender, GridViewCommandEventArgs e)
         {
-            if (e.CommandName == "add")
+            if (e.CommandName == "insert")
             {
                 string RoleName = ((TextBox)gvRoles.FooterRow.FindControl("txtRoleName")).Text;
-                AddNew(RoleName);
-            }
-        }
-
-        protected void AddNew(string RoleName)
-        {
-            if (!Roles.RoleExists(RoleName))
-            {
-                Roles.CreateRole(RoleName);
-                BindData();
+                if (!Roles.RoleExists(RoleName))
+                {
+                    Roles.CreateRole(RoleName);
+                    BindData();
+                }
             }
         }
 
         protected void gvRoles_RowDeleting(object sender, GridViewDeleteEventArgs e)
         {
             string RoleName = ((Label)gvRoles.Rows[e.RowIndex].Cells[1].FindControl("lblRoleName")).Text;
-            Roles.DeleteRole(RoleName);
+            if (RoleName != "host" && RoleName != "admin")
+            {
+                Roles.DeleteRole(RoleName);
+                BindData();
+            }
+        }
+
+        protected void gvRoles_RowEditing(object sender, GridViewEditEventArgs e)
+        {
+            gvRoles.EditIndex = e.NewEditIndex;
             BindData();
         }
 
-        protected void gvRoles_RowDataBound(object sender, GridViewRowEventArgs e)
+        protected void gvRoles_RowCancelingEdit(object sender, GridViewCancelEditEventArgs e)
         {
-            if (e.Row.RowType == DataControlRowType.DataRow)
-            {
-                foreach (ImageButton button in e.Row.Cells[0].Controls.OfType<ImageButton>())
-                {
-                    if (button.CommandName == "Delete")
-                    {
-                        button.Attributes["onclick"] = "if(!confirm('Do you want to delete?')){ return false; };";
-                    }
-                }
-            }
+            gvRoles.EditIndex = -1;
+            BindData();
         }
+
+        //protected void gvRoles_RowDataBound(object sender, GridViewRowEventArgs e)
+        //{
+        //    if (e.Row.RowType == DataControlRowType.DataRow)
+        //    {
+        //        foreach (ImageButton button in e.Row.Cells[0].Controls.OfType<ImageButton>())
+        //        {
+        //            if (button.CommandName == "Delete")
+        //            {
+        //                button.Attributes["onclick"] = "if(!confirm('Do you want to delete?')){ return false; };";
+        //            }
+        //        }
+        //    }
+        //}
     }
 }
